Fixed some more bugs, did some housekeeping on the code
Cleaned up some bugs that were not allowing me to double click on
the same room or change users without losing the messages from the
already loaded room.

Also fixed a bug that would not clear messages when a room without
messages would be chosen, as well as some general housekeeping.